Made it possible to laod examples rather than run tests
